+++ Helpers/Global.cs
@@ -1,0 +1,14 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Web;
+
+namespace WebApplication.Helpers
+{
+    public static class Global
+    {
+
+        public static string ERROR_APPLICATION = "99";
+
+    }
+}